backend/requirements.txt
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
RPi.GPIO==0.7.1
six==1.16.0
uWSGI==2.0.20
Werkzeug==2.1.2
zipp==3.8.0
